+++ CHANGELOG.md
@@ -38,6 +38,7 @@
 - Changed electrolyzer capacity to be specified as `n_clusters` rather than `rating` in electrolyzer performance model config
 - Changed electrolyzer capacity to be an input to the electrolyzer cost models rather than pulled from the cost model config.
 - Updated finance parameter organization naming in `plant_config`.
+- Add ocean alkalinity enhancement technology model.
 
 
 ## 0.3.0 [May 2 2025]